new_input_handling/requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
openai
python-dotenv